Completed Evolution Mechanic

+++ Assets/Scripts/Formula/AbilityFormula.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public abstract class AbilityFormula : ScriptableObject
 {
     public float baseValue;
-    public int multiplier;
+    public float multiplier;
 
     public abstract float Evaluate(int level);
 }
